refactor key
